Enum mpv::Event

pub enum Event<'a> {
    Shutdown,
    LogMessage {
        prefix: &'static str,
        level: &'static str,
        text: &'static str,
        log_level: LogLevel,
    },
    GetPropertyReply {
        name: &'static str,
        result: Result<Format<'a>>,
        reply_userdata: u32,
    },
    SetPropertyReply(Result<()>, u32),
    CommandReply(Result<()>, u32),
    StartFile,
    EndFile(Result<EndFileReason>),
    FileLoaded,
    TracksChanged,
    TrackSwitched,
    Idle,
    Pause,
    Unpause,
    Tick,
    ClientMessage,
    VideoReconfig,
    AudioReconfig,
    MetadataUpdate,
    Seek,
    PlaybackRestart,
    PropertyChange {
        name: &'static str,
        change: Format<'a>,
        reply_userdata: u32,
    },
    ChapterChange,
    QueueOverflow,
    Unused,
}

Variants
Shutdown

Received when the player is shutting down

LogMessage

Has not been tested, received when explicitly asked to MPV

Fields

prefix: &'static str

level: &'static str

text: &'static str

log_level: LogLevel

GetPropertyReply

Received when using get_property_async

Fields

name: &'static str

result: Result<Format<'a>>

reply_userdata: u32

SetPropertyReply(Result<()>, u32)

Received when using set_property_async

CommandReply(Result<()>, u32)

Received when using command_async

StartFile

Event received when a new file is playing

EndFile(Result<EndFileReason>)

Event received when the file being played currently has stopped, for an error or not

FileLoaded

Event received when a file has been loaded, but has not been started

TracksChangedTrackSwitched

Deprecated

Idle

Received when the player has no more files to play and is in an idle state

Pause

The player paused playback

Unpause

The player started playback again

TickClientMessage

Unimplemented

VideoReconfigAudioReconfigMetadataUpdateSeek

The player changed current position

PlaybackRestartPropertyChange

Received when used with observe_property

Fields

name: &'static str

change: Format<'a>

reply_userdata: u32

ChapterChangeQueueOverflow

Received when the Event Queue is full

Unused

Unused event