Struct mpv::MpvHandlerWithGl

pub struct MpvHandlerWithGl {
    // some fields omitted
}

This struct is a decorator of MpvHandler, and can use all the functions from MpvHandler. It is only used when you must embed mpv somewhere else using openGL.

Methods

impl MpvHandlerWithGl

fn draw(&mut self, fbo: i32, width: i32, heigth: i32) -> Result<()>

Render video

The video will use the full provided framebuffer. Options like "panscan" are applied to determine which part of the video should be visible and how the video should be scaled. You can change these options at runtime by using the mpv property API.

fbo is the framebuffer object to render on. Because the renderer might manage multiple FBOs internally for the purpose of video postprocessing, it will always bind and unbind FBOs itself. If you want mpv to render on the main framebuffer, pass 0.

width is the width of the framebuffer. This is either the video size if the fbo parameter is 0, or the allocated size of the texture backing the fbo. The renderer will always use the full size of the fbo.

height is the height of the framebuffer. Same as with the w parameter, except that this parameter can be negative. In this case, the video frame will be rendered flipped.

Errors

If the external video module has not been configured correctly, libmpv can send various errors such as MPV_ERROR_UNSUPPORTED

fn is_update_available(&self) -> bool

returns true if another frame is available